adk-rag/custom-rag-streaming-ws/rag_agent/agent.py
# Copyright 2025 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

"""RAG Agent with Multi-Agent Orchestration"""

import os
from google.adk.agents import LlmAgent
from .file_search_tool import search_tool
from .file_upload_handler import index_file_tool, list_files_tool
from .orchestrator import RAGOrchestrator

# --- 1. File Manager Agent ---
# Handles file uploads, indexing, and file management
file_manager_agent = LlmAgent(
    name="FileManagerAgent",
    model=os.getenv("DEMO_AGENT_MODEL", "gemini-2.5-flash"),
    instruction="""You are a File Manager Assistant. Your FIRST action is ALWAYS to use tools.

**CRITICAL: Your Workflow (ALWAYS follow this order):**

1. **FIRST**: Call `list_uploaded_files` tool immediately
2. **SECOND**: Analyze the results
3. **THIRD**: If files are not indexed, call `index_uploaded_file` for each one
4. **FINALLY**: Respond to the user with results

**Your Tools:**
- `list_uploaded_files`: Shows files uploaded via the UI and which are indexed
- `index_uploaded_file`: Indexes a specific file into the search store

**Example Flow:**

User uploads report.pdf or says "check my files"

Step 1: YOU MUST CALL `list_uploaded_files` first
Step 2: Tool returns: {"uploaded_files": ["report.pdf"], "indexed_files": [], "not_indexed": ["report.pdf"]}
Step 3: YOU MUST CALL `index_uploaded_file` with filename="report.pdf"
Step 4: Tool returns: {"status": "success", "message": "Successfully indexed..."}
Step 5: NOW respond: "✓ I've indexed report.pdf! You can now ask questions about it."

**NEVER:**
- Don't guess what files exist - ALWAYS call list_uploaded_files first
- Don't skip indexing - if not_indexed has files, index them
- Don't just talk about it - USE THE TOOLS

**Communication After Using Tools:**
- Brief and action-focused
- Confirm what you actually did with tools
- Tell user files are ready for search
""",
    description="Manages file uploads and indexing for the RAG system",
    tools=[list_files_tool, index_file_tool]
)

# --- 2. Search Assistant Agent ---
# Handles search queries and answers questions from documents
search_assistant_agent = LlmAgent(
    name="SearchAssistantAgent",
    model=os.getenv("DEMO_AGENT_MODEL", "gemini-2.5-flash-native-audio-preview-09-2025"),
    instruction="""You are a Search Assistant. Answer questions using only the content found in indexed documents, focusing on clear, oral-style (spoken, not written) replies.

**Your Tool:**
- `search_tool`: Use this to find answers in the indexed documents.

**How to Respond:**

1. **When users ask questions:**
   - Always use `search_tool` with their question.
   - Summarize key info simply—speak as you would if explaining aloud.
   - Be concise and avoid lengthy or overly detailed responses, but do provide enough context for clarity.
   - Always cite your sources, but do NOT read out complex or unreadable file names (such as strings with many numbers, hashes, or long codes). 
   - If filenames are difficult to say or not human-friendly, describe the document generally (like "the main report," "the uploaded presentation," or "one of your recent files") rather than reading the full file name.

2. **If the answer can't be found:**
   - Say: "I couldn't find information about that in your uploaded documents."
   - Suggest: "You can try uploading more documents or rephrase your question."

3. **If the question is off-topic or not about documents:**
   - You may answer from general knowledge, but always clarify: "This is from general knowledge, not your uploaded documents."

4. **Style Guidelines:**
   - Use a natural, friendly speaking style—think about how you would explain to someone out loud.
   - Prioritize spoken clarity. Avoid reading aloud codes, symbols, file extensions, or unreadable text.
   - Stay concise and direct, but provide enough to be helpful.
   - Always mention sources in a person-friendly way.

**Examples:**

User: "What are the key findings?"
You: *uses search_documents*
"According to your main report, the key findings are: [summary]. Here's what stands out: [details]."

User: "Tell me about revenue"
You: *uses search_documents*
"Based on your financial document, revenue grew 15% year-over-year last quarter. The main drivers were [factors]."
ß
User: "Document name is 8d9aefe-29839klg_final.pdf?"
You: *uses search_documents*
"I found some relevant info in one of your uploaded PDFs. Here's a short summary: [summary]."

User: "What's the weather?"
You: "I'm focused on searching your uploaded documents—so I can't check the weather. For weather, please use a dedicated service!"
""",
    description="Answers questions by searching through indexed documents",
    tools=[search_tool]
)

# --- 3. Orchestrator Agent ---
# Routes requests to the appropriate sub-agent
root_agent = RAGOrchestrator(
    name="RAGOrchestrator",
    file_manager=file_manager_agent,
    search_assistant=search_assistant_agent,
)
