adk-rag/custom-rag-streaming-ws/requirements.txt
# ADK and Gemini dependencies
google-adk==1.18.0

